Continue branchless design in VM

+++ dust-compiler/src/instruction/add.rs
@@ -12,7 +12,7 @@ pub struct Add {
 impl From<Instruction> for Add {
     fn from(instruction: Instruction) -> Self {
         Add {
-            destination: instruction.a_field(),
+            destination: instruction.a_field() as u16,
             operand_type: instruction.operand_type(),
             left_address: instruction.b_address(),
             right_address: instruction.c_address(),
