+++ dust-compiler/src/instruction/multiply.rs
@@ -12,7 +12,7 @@ pub struct Multiply {
 impl From<Instruction> for Multiply {
     fn from(instruction: Instruction) -> Self {
         Multiply {
-            destination: instruction.a_field(),
+            destination: instruction.a_field() as u16,
             operand_type: instruction.operand_type(),
             left_address: instruction.b_address(),
             right_address: instruction.c_address(),
